+++ libsolidity/ast/AST.h
@@ -218,16 +218,15 @@ public:
 	Visibility visibility() const { return m_visibility == Visibility::Default ? defaultVisibility() : m_visibility; }
 	bool isPublic() const { return visibility() >= Visibility::Public; }
 	virtual bool isVisibleInContract() const { return visibility() != Visibility::External; }
-	bool isVisibleInDerivedContracts() const { return isVisibleInContract() && visibility() >= Visibility::Internal; }
+	virtual bool isVisibleInDerivedContracts() const { return isVisibleInContract() && visibility() >= Visibility::Internal; }
 	bool isVisibleAsLibraryMember() const { return visibility() >= Visibility::Internal; }
+	virtual bool isVisibleViaContractName() const { return false; }
 
 
 	virtual bool isLValue() const { return false; }
 	virtual bool isPartOfExternalInterface() const { return false; }
 
 	/// @returns the type of expressions referencing this declaration.
-	/// The current contract has to be given since this context can change the type, especially of
-	/// contract types.
 	/// This can only be called once types of variable declarations have already been resolved.
 	virtual TypePointer type() const = 0;
 
@@ -418,13 +417,16 @@ public:
 	bool isInterface() const { return m_contractKind == ContractKind::Interface; }
 	bool isLibrary() const { return m_contractKind == ContractKind::Library; }
 
+	/// @returns true, if the contract derives from @arg _base.
+	bool derives(ContractDefinition const& _base) const;
+
 	/// @returns a map of canonical function signatures to FunctionDefinitions
 	/// as intended for use by the ABI.
 	std::map<util::FixedHash<4>, FunctionTypePointer> interfaceFunctions() const;
 	std::vector<std::pair<util::FixedHash<4>, FunctionTypePointer>> const& interfaceFunctionList() const;
 
-	/// @returns a list of the inheritable members of this contract
-	std::vector<Declaration const*> const& inheritableMembers() const;
+	/// @returns a list of all declarations in this contract
+	std::vector<Declaration const*> declarations() const { return filteredNodes<Declaration>(m_subNodes); }
 
 	/// Returns the constructor or nullptr if no constructor was specified.
 	FunctionDefinition const* constructor() const;
@@ -458,7 +460,6 @@ private:
 
 	mutable std::unique_ptr<std::vector<std::pair<util::FixedHash<4>, FunctionTypePointer>>> m_interfaceFunctionList;
 	mutable std::unique_ptr<std::vector<EventDefinition const*>> m_interfaceEvents;
-	mutable std::unique_ptr<std::vector<Declaration const*>> m_inheritableMembers;
 };
 
 class InheritanceSpecifier: public ASTNode
@@ -529,6 +530,9 @@ public:
 
 	TypePointer type() const override;
 
+	bool isVisibleInDerivedContracts() const override { return true; }
+	bool isVisibleViaContractName() const override { return true; }
+
 	TypeDeclarationAnnotation& annotation() const override;
 
 private:
@@ -546,6 +550,9 @@ public:
 		Declaration(_location, _name), m_members(_members) {}
 	void accept(ASTVisitor& _visitor) override;
 	void accept(ASTConstVisitor& _visitor) const override;
+
+	bool isVisibleInDerivedContracts() const override { return true; }
+	bool isVisibleViaContractName() const override { return true; }
 
 	std::vector<ASTPointer<EnumValue>> const& members() const { return m_members; }
 
@@ -704,6 +711,10 @@ public:
 	bool isVisibleInContract() const override
 	{
 		return Declaration::isVisibleInContract() && isOrdinary();
+	}
+	bool isVisibleViaContractName() const override
+	{
+		return visibility() >= Visibility::Public;
 	}
 	bool isPartOfExternalInterface() const override { return isPublic() && isOrdinary(); }
 
@@ -868,6 +879,8 @@ public:
 
 	TypePointer type() const override;
 
+	Visibility defaultVisibility() const override { return Visibility::Internal; }
+
 	ModifierDefinitionAnnotation& annotation() const override;
 
 private:
@@ -927,6 +940,9 @@ public:
 
 	TypePointer type() const override;
 	FunctionTypePointer functionType(bool /*_internal*/) const override;
+
+	bool isVisibleInDerivedContracts() const override { return true; }
+	bool isVisibleViaContractName() const override { return false; /* TODO */ }
 
 	EventDefinitionAnnotation& annotation() const override;